refactor: formatter 패키지 삭제
