+++ src/main/java/domain/LottoDispenser.java
@@ -1,6 +1,5 @@
 package domain;
 
-import domain.formatter.LottoBuyResultFormatter;
 import exception.LottoException;
 import java.util.ArrayList;
 import java.util.List;
