+++ src/main/java/domain/WinningCase.java
@@ -1,6 +1,5 @@
 package domain;
 
-import domain.formatter.WinningCaseFormatter;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
