+++ src/main/java/service/MockingLottoService.java
@@ -5,7 +5,7 @@ import domain.Lotto;
 import domain.LottoDispenser;
 import domain.WinningCase;
 import domain.WinningNumber;
-import domain.formatter.WinningCalculateFormatter;
+import domain.WinningCalculateFormatter;
 import java.util.List;
 import java.util.Map;
 import repository.BonusNumberRepository;
